Correct linting

+++ app/core/background_tasks.py
@@ -1,5 +1,6 @@
 import os
 
 
-def cleanup_file(file_path: str):
-    os.remove(file_path)
+def cleanup_file(file_path: str) -> None:
+    if os.path.exists(file_path):
+        os.remove(file_path)
